Comments+SelectionSort+abstractSort class
Large patch, contains an update adding in an abstract sort class that contains the comarable array that we use in the sort classes.

Added a large number of comments explaining what the code is doing along with the appropriate links within the code to help readability.

Selection sort is complete and working along with the tests.

Updated the tests as well to add a line of logging so we can see the output from the arrays.

Missed an = sign in the implmentation that had me scratching my head for a few minutes until I stepped through. LOL!