Added Main Fragment with viewmodel.
There is alot of code that needs commenting, Will work on that if i have the time. Did not get started till nearly midnight.

There is a view model for the main activity, main fragment, sortfragment.

Need to create the intermediate fragment for the list of algorithms we can use.

The sort fragment will take in the different types of sort and then we will generate the data using the generate data method. Once that is complete need to work on the visualizer.

Added the architecture components to this commit as well since Viewmodel and Live data are part of that.

LiveData is pretty useful. It may be useful in removing the SortHighlighter interface.

Note there is a lot of code that may get removed or modified in this commit. Coming back to this after a while. Taking some time to get back in the groove.